backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers
django-filter
psycopg2-binary
redis
celery[redis]
PyYAML
Pillow
requests
markdown
bleach
python-dotenv
gunicorn
whitenoise
django-extensions
PyJWT
cryptography
openpyxl
lxml
